🚚 Trekk endreVilkårPerioder ut i egen fil
